graph: make load() private
This is now called by the constructor.
To load a new/different graph you have to instantiate a new object now.
This reduces possible side effects, ArrayRepresentation is now
effectively free of side effects (once the graph is loaded) and all
getters can be accessed without the need for locks or mutual exclusion
(except of calls relayed to LookupGrid).